Change project structure to be more modular.
